% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_met_files.R
\name{download_met_files}
\alias{download_met_files}
\title{Download required met files}
\usage{
download_met_files(
  met_type,
  days,
  duration = 0,
  direction = "forward",
  met_dir
)
}
\arguments{
\item{met_type}{Type of met file.}

\item{days}{Vector of dates to download.}

\item{duration}{Duration of eventual HYSPLIT runs (determines whether extra dates ar necessary)}

\item{direction}{Direction of eventual HYSPLIT runs; "forward" or "backward".}

\item{met_dir}{Directory into which met files should be downloaded/checked for.}
}
\description{
Download required met files
}
