% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_dispersion_execute.R
\name{hysplit_dispersion_execute}
\alias{hysplit_dispersion_execute}
\title{Conduct HYSPLIT dispersion runs}
\usage{
hysplit_dispersion_execute(disp_setup)
}
\arguments{
\item{disp_setup}{List output from hysplit_dispersion.R}
}
\description{
The function executes single/multiple forward or backward HYSPLIT dispersion
runs using specified meteorological datasets.
}
