% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the model}
\usage{
run_model(model, particle_max = "100000")
}
\arguments{
\item{model}{A splitr modeling object.}
}
\description{
Run either a trajectory model or a dispersion model, depending on the class
of the splitr modeling object.
}
