% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_dispersion.R
\name{hysplit_dispersion}
\alias{hysplit_dispersion}
\title{Conduct HYSPLIT dispersion runs}
\usage{
hysplit_dispersion(
  geo_df,
  start_day = "2015-07-01",
  start_hour = 0,
  duration = 24,
  direction = "forward",
  met_type = "reanalysis",
  vert_motion = 0,
  model_height = 20000,
  particle_num = 10,
  particle_max = 100,
  species,
  disp_name = NULL,
  binary_path = NULL,
  binary_name = NULL,
  exec_dir = NULL,
  met_dir = NULL,
  softrun = NULL,
  clean_up = TRUE
)
}
\arguments{
\item{geo_df}{A data frame with columns "lat", "lon", and "height" denoting starting points.}

\item{start_day}{the day that the model will initialize and run. This should
take the form of a single-length vector for a day (\code{"YYYY-MM-DD"}).}

\item{start_hour}{a single daily hour as an integer hour (from \code{0} to \code{23}).}

\item{duration}{The duration of each model run (whether it is in the forward
direction or running backwards) in hours.}

\item{direction}{An option to select whether to conduct the model in the
\code{"forward"} (default) or \code{"backward"} directions.}

\item{met_type}{The type of meteorological data files to use. The options
are: \code{"reanalysis"} (NCAR/NCEP global reanalysis data, the default),
\code{"gdas1"} and \code{"gdas0.5"} (Global Data Assimilation System 1-degree and
0.5-degree resolution data), \code{"narr"} (North American Regional Reanalysis),
\code{"gfs0.25"} (Global Forecast System 0.25 degree data), and \code{"nam12"} (North
American Mesoscale Forecast System, 12-km/6-hour resolution data).}

\item{vert_motion}{A numbered option to select the method used to simulation
vertical motion. The methods are: \code{0} (input model data), \code{1} (isobaric),
\code{2} (isentropic), \code{3} (constant density), \code{4} (isosigma), \code{5} (from
divergence), \code{6} (remap MSL to AGL), \code{7} (average data), and \code{8} (damped
magnitude).}

\item{model_height}{The upper limit of the model domain in meters.}

\item{particle_num}{The number of particles released by a source during each
release cycle.}

\item{particle_max}{The maximum number of particles released by a source
during a model run.}

\item{species}{A list of parameters corresponding to an emissions source..}

\item{disp_name}{An optional, descriptive name for the output file
collection.}

\item{binary_path}{An optional path to a HYSPLIT trajectory model binary.
When not specified, the model binary will be chosen from several available
in the package (based on the user's platform).}

\item{binary_name}{An optional file name for the hysplit binary code. Defaults to "hycs_std", but could be "hycm_std", for example.}

\item{exec_dir}{An optional file path for the working directory of the model
input and output files.}

\item{met_dir}{An optional file path for storage and access of meteorological
data files.}

\item{clean_up}{An option to make the \code{exec_dir} directory clean after
completion of all trajectory runs. By default, this is set to \code{TRUE}.}
}
\description{
The function executes single/multiple forward or backward HYSPLIT dispersion
runs using specified meteorological datasets.
}
