% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysplit_multitraj.R
\name{hysplit_multitraj}
\alias{hysplit_multitraj}
\title{Conduct extract temperature data from multiple starting points in HYSPLIT}
\usage{
hysplit_multitraj(
  full_df,
  direction = "forward",
  met_type = "nam12",
  vert_motion = 0,
  model_height = 20000,
  extended_met = TRUE,
  temp_only = TRUE,
  traj_name = NULL,
  met_dir = NULL,
  exec_dir = NULL,
  system_type = "win"
)
}
\arguments{
\item{direction}{An option to select whether to conduct the model in the
\code{"forward"} (default) or \code{"backward"} directions.}

\item{met_type}{The type of meteorological data files to use. The options
are: \code{"reanalysis"} (NCAR/NCEP global reanalysis data, the default),
\code{"gdas1"} and \code{"gdas0.5"} (Global Data Assimilation System 1-degree and
0.5-degree resolution data), \code{"narr"} (North American Regional Reanalysis),
\code{"gfs0.25"} (Global Forecast System 0.25 degree data), and \code{"nam12"} (North
American Mesoscale Forecast System, 12-km/6-hour resolution data).}

\item{vert_motion}{A numbered option to select the method used to simulation
vertical motion. The methods are: \code{0} (input model data), \code{1} (isobaric),
\code{2} (isentropic), \code{3} (constant density), \code{4} (isosigma), \code{5} (from
divergence), \code{6} (remap MSL to AGL), \code{7} (average data), and \code{8} (damped
magnitude).}

\item{model_height}{The upper limit of the model domain in meters.}

\item{extended_met}{An option to report additional meteorological data along
each output trajectory.}

\item{traj_name}{An optional, descriptive name for the output file
collection.}

\item{met_dir}{An optional file path for storage and access of meteorological
data files.}

\item{exec_dir}{An optional file path for the working directory of the model
input and output files.}
}
\description{
Conduct extract temperature data from multiple starting points in HYSPLIT
}
